requirements.txt
pytest
pytest-redis
requests
factory_boy